refactored classic annealing schedule tests
